backend/requirements.txt
crawl4ai==0.6.0
langchain==0.1.20
langchain-openai
langchain-pinecone==0.0.3
pinecone-client==3.0.2
streamlit==1.29.0
python-dotenv==1.0.0
requests==2.31.0
tiktoken==0.7.
